+++ packages/design-core/index.js
@@ -33,7 +33,7 @@ export { default as Styles } from '@opentiny/tiny-engine-setting-styles'
 export { default as Layout, LayoutService } from '@opentiny/tiny-engine-layout'
 export { default as Canvas } from '@opentiny/tiny-engine-canvas'
 export { initPreview } from './src/preview/src/main'
-export { GenerateCodeService, PluginPanel } from '@opentiny/tiny-engine-common'
+export { GenerateCodeService, PluginPanel, PluginSetting } from '@opentiny/tiny-engine-common'
 
 export { default as defaultRegistry } from './registry'
