+++ packages/plugins/page/index.js
@@ -13,11 +13,17 @@
 import entry, { api } from './src/Main.vue'
 import metaData from './meta.js'
 import { PageService } from './src/composable'
+import PageGeneral from './src/PageGeneral.vue'
+import PageMain from './src/PageMain.vue'
 
 export default {
   ...metaData,
   apis: api,
   entry,
+  components: {
+    PageMain,
+    PageGeneral
+  },
   metas: [PageService]
 }
