+++ packages/plugins/page/src/PageFolderSetting.vue
@@ -18,7 +18,7 @@
       <div class="page-setting-content">
         <tiny-collapse v-model="state.activeName">
           <tiny-collapse-item title="基本设置" name="folderGeneralRef">
-            <page-general ref="folderGeneralRef" :isFolder="isFolder"></page-general>
+            <component :is="pageGeneral" ref="folderGeneralRef" :isFolder="isFolder"></component>
           </tiny-collapse-item>
         </tiny-collapse>
       </div>
@@ -30,10 +30,10 @@
 import { reactive, ref } from 'vue'
 import { Button, Collapse, CollapseItem } from '@opentiny/vue'
 import { PluginSetting, SvgButton, ButtonGroup } from '@opentiny/tiny-engine-common'
-import { usePage, useModal, useApp, useNotify } from '@opentiny/tiny-engine-meta-register'
+import { usePage, useModal, useApp, useNotify, getMergeRegistry } from '@opentiny/tiny-engine-meta-register'
 import { isEqual } from '@opentiny/vue-renderless/common/object'
 import throttle from '@opentiny/vue-renderless/common/deps/throttle'
-import PageGeneral from './PageGeneral.vue'
+import meta from '../meta'
 import http from './http.js'
 
 let isShow = ref(false)
@@ -55,7 +55,6 @@ export default {
     TinyCollapse: Collapse,
     TinyCollapseItem: CollapseItem,
     PluginSetting,
-    PageGeneral,
     SvgButton,
     ButtonGroup
   },
@@ -70,11 +69,13 @@ export default {
       activeName: ['folderGeneralRef'],
       title: '文件夹设置'
     })
-    const folderGeneralRef = ref(null)
     const { requestCreatePage, requestUpdatePage, requestDeletePage } = http
     const { appInfoState } = useApp()
     const { pageSettingState, changeTreeData } = usePage()
     const { confirm } = useModal()
+    const registry = getMergeRegistry(meta.type, meta.id)
+    const pageGeneral = registry.components.PageGeneral
+    const folderGeneralRef = ref(null)
 
     const closeFolderSetting = () => {
       if (isEqual(pageSettingState.currentPageData, pageSettingState.currentPageDataCopy)) {
@@ -192,6 +193,7 @@ export default {
     return {
       saveFolderSetting,
       deleteFolder: throttle(5000, true, deleteFolder),
+      pageGeneral,
       folderGeneralRef,
       closeFolderSettingPanel,
       isShow,
